Add files via upload
First commit - This file is an optimized energy calculation of a 25x25 randomly generated lattice, comprised of up and down spins (represented as 1 and 0 in the matrix).  It correctly calculates the energy of any given lattice, but we still have more work to do on getting the Monte Carlo algorithm to function correctly.